+++ ADMF/changelog.md
@@ -3,6 +3,7 @@
 ## ???
 
 - Upd: Set-AdmfContext - added Context name insertion for forest Components that support it.
+- Fix: Fails to resolve a domain controller using a credential provider due to timing issues.
 
 ## 1.1.23 (2020-04-17)